// src/app/shared/shared/clinic-modules.enum.ts
export class ClinicModulesEnum {

    Coding: string = "Coding";
    ClinicMaster: string = "ClinicMaster";
    AccessRights: string = "AccessRights";
    PatientRegistration: string = "PatientRegistration";
    Appointments: string = "Appointments";
    EMR: string = "EMR";
    BillingAndInsurance: string = "BillingAndInsurance";
    PurchasingAndInventory: string = "PurchasingAndInventory";

}